+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - AppImage
     - able to update AppImages with continuous releases
 - UI
-    - not performing a full table refresh after installing a new package
+    - not performing a full table refresh after installing and uninstalling packages
     - filters algorithm speed and sorting
     - displaying wait cursor over some components